Save registry data per-connection
Also disables registry on local connection
